LootingSystem
- LootingSystem basic work
- need to work on cardSelectUI next
