+++ Assets/Contents/Sprites/Icons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab4ce7f88f68723409447120d8165182
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
