+++ Assets/Scripts/Common/Game/LootingItems.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 [CreateAssetMenu(fileName = "LootingCard", menuName = "Scriptable Objects/Looting/LootingCardDataSO")]
 [System.Serializable]
-public class LootingCard : ScriptableObject
+public class LootingCardSO : ScriptableObject
 {
     public int lootingStage;
     public int lootPower; // 1~100;
@@ -11,14 +11,14 @@ public class LootingCard : ScriptableObject
 
 [CreateAssetMenu(fileName = "LootingPotion", menuName = "Scriptable Objects/Looting/LootingPotionDataSO")]
 [System.Serializable]
-public class LootingPotion : ScriptableObject
+public class LootingPotionSO : ScriptableObject
 {
     public int lootingStage;
     public int lootPower;
     public PotionDataSO potionDataSO;
 }
 
-public class LootingTreature : ScriptableObject
+public class LootingTreasureSO : ScriptableObject
 {
     public int lootingStage;
     public int lootPower;
