+++ Assets/Scripts/Common/UI/BattleUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 92fb2d14b2da14c4faef83f9afd1ca47
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
