+++ Assets/Scripts/Common/UI/BattleUI/DiceRollUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 041138ff976dd894d9be182e63a2806d
