+++ Assets/Scripts/Common/UI/BattleUI/PotionSlotUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 97a268f51d3247e44ad771b650dfe177
