+++ Assets/Scripts/Common/UI/BattleUI/SkillDiceSlotUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 99fb0ba12c104af4ca2a3f5192b20a9b
