+++ Assets/Scripts/Common/UI/BattleUI/SkillUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9a71e1c9d221493458be84e08c2dec97
