+++ Assets/Scripts/Common/UI/BattleUI/SkillUIInteract.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4d63fa76b287a1f4aa6e45300a06cce3
