--- Assets/Scripts/Common/UI/DiceRollUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 041138ff976dd894d9be182e63a2806d
