+++ Assets/Scripts/Common/UI/LootngUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 28c8eabf7b17ef946ae82a5fd22b3834
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
