+++ Assets/Scripts/Common/UI/LootngUI/CardSelectUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5636ac209b814d84d9ece8317a5eecb1
