+++ Assets/Scripts/Common/UI/LootngUI/LootingDataBase.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+[System.Serializable]
+public class LootingDataBase
+{
+    // ÀÏ¹Ý Looting È®·ü
+    public LootingProbTable NormalBattleLoot;
+    public LootingProbTable EliteBattleLoot;
+    public LootingProbTable BossBattleLoot;
+    public LootingProbTable TreasureLoot;
+
+    public Sprite LootingCardImage;
+}
