+++ Assets/Scripts/Common/UI/LootngUI/LootingDataBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: eed4cce9cabd4534e953d4ccb82feb8f
