+++ Assets/Scripts/Common/UI/LootngUI/LootingItemUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2a2ffb3337ff80144bfce22c37eb1a6c
