+++ Assets/Scripts/Common/UI/LootngUI/LootingUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8dc70b5286310084585e677c3b83bb09
