+++ Assets/Scripts/Common/UI/LootngUI/ShopItemUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2646fbdc0e1da384fac5c91a8c1e69f2
