+++ Assets/Scripts/Common/UI/LootngUI/ShopUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0dca9543224b7e646a697eec8856c9ec
