--- Assets/Scripts/Common/UI/PotionSlotUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 97a268f51d3247e44ad771b650dfe177
