--- Assets/Scripts/Common/UI/ShopItemUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2646fbdc0e1da384fac5c91a8c1e69f2
