--- Assets/Scripts/Common/UI/SkillDiceSlotUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 99fb0ba12c104af4ca2a3f5192b20a9b
