--- Assets/Scripts/Common/UI/SkillUIInteract.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4d63fa76b287a1f4aa6e45300a06cce3
