--- Assets/Scripts/Common/UI/SkillUISlot.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 43fc048b3ced6894b9a016e36e20a378
